pyproject.toml
[project]
name = "studymathai"
version = "2.0"
description = ""
authors = [
    {name = "Ruskin Patel",email = "ruskin.patel23@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.82.0,<2.0.0)",
    "chromadb (>=1.0.10,<2.0.0)",
    "pymupdf (>=1.26.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (==0.115.9)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "studymathai", from = "src"}]


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
pypdf2 = "^3.0.1"
pytest = "^8.4.1"
ruff = "^0.6.9"
black = "^24.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
  ".git",
  "__pycache__",
  ".venv",
  "venv",
  "build",
  "dist",
  "chroma_index",
]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "W",  # pycodestyle warnings
  "I",  # isort (import sorting)
  "B",  # flake8-bugbear
  "UP", # pyupgrade
]
ignore = [
]

[tool.ruff.lint.isort]
known-first-party = ["studymathai", "api"]
combine-as-imports = true

[tool.black]
target-version = ["py312"]
line-length = 100
